Add files via upload
This is the source code for the TodoList app which I can type in any item in the text box and then I press a button to add the item. From there, the item gets added underneath the button along with the check box on the left of the name of the item and the delete button on the right of the name of the item. I can also keep on adding more items instead of just having only one item.